+++ src/Service/CognitoIdentityProvider/src/Exception/ForbiddenException.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace AsyncAws\CognitoIdentityProvider\Exception;
+
+use AsyncAws\Core\Exception\Http\ClientException;
+use Symfony\Contracts\HttpClient\ResponseInterface;
+
+/**
+ * This exception is thrown when WAF doesn't allow your request based on a web ACL that's associated with your user
+ * pool.
+ */
+final class ForbiddenException extends ClientException
+{
+    protected function populateResult(ResponseInterface $response): void
+    {
+        $data = $response->toArray(false);
+
+        if (null !== $v = (isset($data['message']) ? (string) $data['message'] : null)) {
+            $this->message = $v;
+        }
+    }
+}