+++ src/Service/CognitoIdentityProvider/src/Input/SetUserMFAPreferenceRequest.php
@@ -19,7 +19,7 @@ final class SetUserMFAPreferenceRequest extends Input
     private $smsMfaSettings;
 
     /**
-     * The time-based one-time password software token MFA settings.
+     * The time-based one-time password (TOTP) software token MFA settings.
      *
      * @var SoftwareTokenMfaSettingsType|null
      */